rclone/requirements.txt
BeautifulSoup==3.2.1
Django==1.8.4
Django-Select2==5.6.0
Markdown==2.6.5
MarkupSafe==0.23
Pillow==3.0.0
Pygments==2.0.2
Unidecode==0.04.18
Whoosh==2.7.0
amqp==1.4.9
anyjson==0.3.3
app-version==0.2.1
argparse==1.2.1
billiard==3.3.0.22
blinker==1.4
celery==3.1.18
coverage==4.0.3
cssselect==0.9.1
decorator==4.0.6
django-activity-stream==0.6.0
django-appconf==1.0.1
django-disqus==0.5
django-envelope==1.1
django-forms-bootstrap==3.0.1
django-froala-editor==2.0.1
django-guardian==1.3.2
django-haystack==2.4.1
django-honeypot==0.5.0
django-markup==0.5dev
django-realtime==1.1
django-stdimage==2.2.0
django-suit==0.2.15
django-tastypie==0.13.0
django-userena==1.4.1
django-uuslug==1.1.6
docutils==0.12
easy-thumbnails==2.3
enum34==1.1.1
feedgenerator==1.7
funcsigs==0.4
future==0.15.2
goose-extractor==1.0.25
html2text==2014.12.29
jieba==0.38
jsonpath-rw==1.4.0
kombu==3.0.33
lxml==3.5.0
mock==1.3.0
nltk==3.1
pbr==1.8.1
pelican==3.6.3
pilkit==1.1.13
ply==3.8
progressbar2==3.6.0
psycopg2==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-slugify==1.1.4
pytz==2015.7
requests==2.9.1
six==1.10.0
suit==2.0.2
tolerance==0.1.2
wsgiref==0.1.2
